# Conflicts:
#   runnershi_fe/RunnersHi/app/src/main/java/com/A306/runnershi/Fragment/UserSearchFragment.kt

+++ runnershi_be/src/main/java/com/ssafy/runnershi/entity/UserInfo.java
@@ -32,14 +32,14 @@ public class UserInfo implements Serializable {
   private double totalDistance;
   private int totalTime;
   private int totalDay;
-  private int bestPace;
+  private double bestPace;
 
   private double weeklyDistance;
   private int weeklyTime;
-  private int weeklyPace;
+  private double weeklyPace;
 
   private double thisWeekDistance;
   private int thisWeekTime;
-  private int thisWeekPace;
+  private double thisWeekPace;
 
 }
